Refactor tests for utils module
